+++ site/astro.config.mjs
@@ -20,6 +20,7 @@ export default defineConfig({
         Sidebar: './src/components/Sidebar.astro',
         Hero: './src/components/Hero.astro',
         Header: './src/components/Header.astro',
+        PageTitle: './src/components/PageTitle.astro',
         Footer: './src/components/Footer.astro',
       },
       customCss: ['./src/custom.css'],
@@ -63,6 +64,7 @@ export default defineConfig({
         { label: 'About Specs', slug: 'overview/aboutspecs' },
         { label: 'Getting Started', slug: 'cli/getting-started' },
         { label: 'Releases', slug: 'overview/releases' },
+        { label: 'Decision Records (ADRs)', slug: 'adr' },
         { label: 'Licensing', slug: 'overview/licensing', badge: pro },
         {
           label: 'Specs 2 Figma Plugin',
@@ -125,6 +127,7 @@ export default defineConfig({
                 { label: 'props', slug: 'cli/analyze/props' },
                 { label: 'styling', slug: 'cli/analyze/styling' },
                 { label: 'dependencies', slug: 'cli/analyze/dependencies' },
+                { label: 'keys', slug: 'cli/analyze/keys' },
               ],
             },
           ],